refact: separando portas e controllers
