+++ backend/controllers/bus_stop_controller.py
@@ -1,0 +1,15 @@
+from backend.domain.bus_stop.get_stop import GetStop
+from flask import jsonify, render_template
+
+def list_stops():
+    stops_list = GetStop.get_all_stops_port()
+
+    stops_dict = {}
+    for stop in stops_list:
+        stops_dict[stop.stop_id] = {
+            'stop_lat' : stop.stop_lat,
+            'stop_lon': stop.stop_lon,
+            'stop_name' : stop.stop_name
+        }
+        
+    return render_template("stops.html", stops=jsonify(stops_dict))
