+++ backend/domain/user/user_login.py
@@ -1,13 +1,5 @@
-from backend.domain.user.get_user import GetUser
+from backend.domain.user.user_login_impl import UserLoginImpl
 
 class UserLogin:
-
-    def do_login(email, password):
-        user = GetUser.get_user_by_email(email)
-        if user:
-            if user.password == password:
-                return {'message' : 'You are logged in'}
-            else:
-                return {'message' : 'Wrong password'}
-        else:
-            return {'message' : 'User does not exist'}
+    def do_login_port(email, password):
+        return UserLoginImpl.do_login(email, password)
